Your commit message

+++ hw_3/todos/views.py
@@ -1,13 +1,16 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from .models import Todo
 from .forms import TodoForm
+from django.forms.widgets import DateTimeInput
 
-def todo_list(request):
-    todos = Todo.objects.all()
-    return render(request, 'todos/todo_list.html', {'todos': todos})
+# Create your views here.
 
-def todo_detail(request, id):
-    todo = get_object_or_404(Todo, id=id)
+def todos_list(request):
+    todos = Todo.objects.all().order_by('due_date')
+    return render(request, 'todos/todos_list.html', {'todos': todos})
+
+def todo_detail(request, pk):
+    todo = get_object_or_404(Todo, pk=pk)
     return render(request, 'todos/todo_detail.html', {'todo': todo})
 
 def todo_create(request):
@@ -15,12 +18,20 @@ def todo_create(request):
         form = TodoForm(request.POST)
         if form.is_valid():
             form.save()
-            return redirect('todo_list')
+            return redirect('todos_list')
     else:
         form = TodoForm()
     return render(request, 'todos/todo_form.html', {'form': form})
 
-def todo_delete(request, id):
-    todo = get_object_or_404(Todo, id=id)
-    todo.delete()
-    return redirect('todo_list')
+def todo_delete(request, pk):
+    todo = get_object_or_404(Todo, pk=pk)
+    if request.method == 'POST':
+        todo.delete()
+        return redirect('todos_list')
+    return render(request, 'todos/todo_delete.html', {'todo': todo})
+
+def toggle_status(request, pk):
+    todo = get_object_or_404(Todo, pk=pk)
+    todo.status = not todo.status
+    todo.save()
+    return redirect('todos_list')
